Add back support for normal mapping without tangent data
Added a boolean member `m_HasTangentData` to `MeshData` to indicate the presence of tangent data. Updated `Mesh::Initialize` in `Visual.cpp` to set this field based on vertex data. Introduced `CalculateTBNWithoutTangent` in `basepass.hlsl` to compute the TBN matrix when tangent data is missing. Modified `GetGBufferParams` in `basepass.hlsl` to conditionally compute the normal using either the precomputed TBN matrix or the new function. Updated normal calculation to use the appropriate TBN matrix.